+++ apps/explorer/lib/explorer/celo/account_reader.ex
@@ -8,7 +8,7 @@ defmodule Explorer.Celo.AccountReader do
 
   use Bitwise
 
-  import Explorer.Celo.Util
+  import Explorer.Celo.{EpochUtil, Util}
 
   def account_data(%{address: account_address}) do
     data = fetch_account_data(account_address)
@@ -52,6 +52,15 @@ defmodule Explorer.Celo.AccountReader do
     data = call_methods([{:validators, "getValidatorGroup", [address]}])
 
     case data["getValidatorGroup"] do
+      {:ok, [res | _]} -> {:ok, res}
+      _ -> :error
+    end
+  end
+
+  def get_carbon_offsetting_partner(bn) do
+    data = call_methods([{:epochrewards, "carbonOffsettingPartner", [], bn}])
+
+    case data["carbonOffsettingPartner"] do
       {:ok, [res | _]} -> {:ok, res}
       _ -> :error
     end
@@ -239,7 +248,6 @@ defmodule Explorer.Celo.AccountReader do
   end
 
   def active_votes(%{
-        block_hash: block_hash,
         block_number: block_number,
         group_hash: group_hash,
         account_hash: account_hash
@@ -251,14 +259,7 @@ defmodule Explorer.Celo.AccountReader do
 
     case data["getActiveVotesForGroupByAccount"] do
       {:ok, [active]} ->
-        {:ok,
-         %{
-           account_hash: account_hash,
-           active_votes: active,
-           block_hash: block_hash,
-           block_number: block_number,
-           group_hash: group_hash
-         }}
+        {:ok, active}
 
       _ ->
         :error