+++ apps/explorer/test/support/factory.ex
@@ -22,6 +22,7 @@ defmodule Explorer.Factory do
     CeloAccount,
     CeloContractEvent,
     CeloCoreContract,
+    CeloElectionRewards,
     CeloPendingEpochOperation,
     CeloValidatorGroupVotes,
     CeloVoterVotes,
@@ -324,8 +325,7 @@ defmodule Explorer.Factory do
       # caller MUST supply block
       # all operations will default to true
       fetch_epoch_rewards: true,
-      fetch_validator_group_data: true,
-      fetch_voter_votes: true
+      election_rewards: true
     }
   end
 
@@ -744,6 +744,17 @@ defmodule Explorer.Factory do
       block_hash: block_hash(),
       block_number: block_number(),
       group_hash: address_hash()
+    }
+  end
+
+  def celo_election_rewards_factory do
+    %CeloElectionRewards{
+      account_hash: address_hash(),
+      amount: Decimal.new(1),
+      associated_account_hash: address_hash(),
+      block_number: block_number(),
+      block_timestamp: DateTime.utc_now(),
+      reward_type: "voter"
     }
   end